+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 [Unreleased]: https://github.com/chaostoolkit/chaostoolkit-lib/compare/1.7.0...HEAD
 
+### Added
+
+- Catch `InterruptExecution` during rollbacks so that the experiment terminates
+  gracefully [#132][132]
+
+[132]: https://github.com/chaostoolkit/chaostoolkit-lib/issues/132
+
 ## [1.7.0][] - 2019-09-21
 
 [1.7.0]: https://github.com/chaostoolkit/chaostoolkit-lib/compare/1.6.0...1.7.0